+++ changelog.md
@@ -26,6 +26,8 @@
 
 *	2020.8.6
 
+	+	staffNotation.ts: fuzzyMatchNotations added.
+
 	+	staffSvg: refined alter token parsing.